refactor to make logo easier to rebrand
