+++ packages/app/src/hooks/index.ts
@@ -4,3 +4,5 @@ export { default as useDerivedColor } from "./use-derived-color";
 export { default as useRedir } from "./use-redir";
 export { default as useUser } from "./use-user";
 export { useWindowSize } from "./use-window-size";
+export { useRouteChange } from "./use-route-change";
+export { useUserWallet } from "./use-user-wallet";
